Download today's Sandmann episode to a local file
Initialized a simple Kotlin project and wrote a simple class that
parses the Sandmann homepage to find a link to the video of the today's
episode. Actually it first finds a JSON descriptor which then leads to
multiple versions of the video from which I choose the one with the highest
quality.